started the cli
