+++ almost_cms/lambda-script.js
@@ -50,16 +50,14 @@ const copyPolicy = (
 );
 
 const trustRel =`{"Version": "2012-10-17","Statement": [{"Effect": "Allow","Principal": {"Service": "lambda.amazonaws.com"},"Action": "sts:AssumeRole"}]}`;
-
 const uniqNow = new Date().toISOString().replace(/-/, '').replace(/-/, '').replace(/T/, '').replace(/\..+/, '').replace(/:/, '').replace(/:/, '');
 console.log(uniqNow);
 
-//CREATE THE IAM POLICY
+// CREATE THE IAM POLICY
 var policyParams = {
     PolicyDocument: copyPolicy, /* required */
     PolicyName: `almostCopyPolicy${uniqNow}`, /* required */
-    Description: 'the IAM policy that allows the role to communicate with your sotrage and public buckets',
-    //Path:'STRING_VALUE'
+    Description: 'the IAM policy that allows the role to communicate with your sotrage and public buckets'
   };
   iam.createPolicy(policyParams, function(err, data) {
     if (err){
@@ -68,7 +66,8 @@ var policyParams = {
     else {
         console.log(data);
         var policyArn = data.Policy.Arn;
-        //CREATE THE IAM ROLE
+        
+        // CREATE THE IAM ROLE
         var rolParams = {
             AssumeRolePolicyDocument: trustRel, /* required */
             RoleName: `almostCopyRole${uniqNow}`, /* required */
@@ -91,7 +90,7 @@ var policyParams = {
             console.log(rolName);
             console.log(rolArn);
 
-            //ATTACH THE IAM POLICY TO THE NEW ROLE
+            // ATTACH THE IAM POLICY TO THE NEW ROLE
             var attachParams = {
                 PolicyArn: policyArn, 
                 RoleName: rolName
@@ -102,10 +101,11 @@ var policyParams = {
                     } 
                     else {
                         console.log(data);
-                        // CREATE THE LAMBDA FUNCTION
+                        
+                        // WAIT 10 SECONDS 
                         setTimeout(
                             function createCopyFunc(){
-
+                                // CREATE THE LAMBDA COPY BUCKET FUNCTION
                                 console.log("ROLE ARN: ", rolArn)
                                 var funcParams = {
                                     Code: {
@@ -135,12 +135,13 @@ var policyParams = {
                                         const copyFuncArn = data.FunctionArn;
                                         const copyFuncName = data.FunctionName;
                                         console.log(copyFuncArn);      
-                                        //CREATE THE API
+                                        
+                                        // CALL THE CREATE API FUNCTION
                                         createApi(copyFuncArn, copyFuncName);
                                         
                                     }
                                 });
-                            }, 10000);
+                            }, 10000); 
                     }     
                });
         }
@@ -149,7 +150,7 @@ var policyParams = {
 });
 
 
-//CREATE THE API
+// CREATE THE API
 function createApi(arn, funcName) {
     var params = {
         name: `almostCopyApi${uniqNow}`, /* required */
@@ -172,116 +173,117 @@ function createApi(arn, funcName) {
             //const rest_api_name = data.name
             console.log("REST API ID: ", rest_api_id)
             
-            //GET PARENT ID
+            // GET THE PARENT RESOURCE ID
             var params = {
                 restApiId: rest_api_id, /* required */
                 };
                 apigateway.getResources(params, function(err, data) {
                 if (err) {
-                    console.log(err, err.stack); // an error occurred
+                    console.log(err, err.stack);
                 }
                 else {
-                    
                     const parent_id = data.items[0].id;
-                    console.log("RESOUCRES: ", parent_id);
-                  
-                            
-                            //CREATE THE POST METHOD
-                            var params = {
-                                authorizationType: 'NONE', /* required */
-                                httpMethod: 'POST', /* required */
-                                resourceId: parent_id, /* required */
-                                restApiId: rest_api_id, /* required */
-                                apiKeyRequired: false,
-                            };
-                            apigateway.putMethod(params, function(err, data) {
-                                if (err) {
-                                    console.log(err, err.stack);
-                                } 
-                                else {
-                                    console.log(data);
-            
-                                    //CREATE THE INTEGRATION WITH THE LAMBDA FUNCTION
-                                    var params = {
-                                        httpMethod: 'POST', /* required */
-                                        integrationHttpMethod: 'POST',
-                                        resourceId: parent_id, /* required */
-                                        restApiId: rest_api_id, /* required */
-                                        type: 'AWS', /* required */
-                                        uri: `arn:aws:apigateway:${process.env.AWS_REGION}:lambda:path/2015-03-31/functions/${arn}/invocations`
-                                    };
-                                    apigateway.putIntegration(params, function(err, data) {
-                                        if (err) {
-                                            console.log(err, err.stack);
-                                        }
-                                        else {
-                                            console.log(data);
+                    console.log("RESOUCRES: ", parent_id);     
+    
+                        // CREATE THE POST METHOD
+                        var params = {
+                            authorizationType: 'NONE', /* required */
+                            httpMethod: 'POST', /* required */
+                            resourceId: parent_id, /* required */
+                            restApiId: rest_api_id, /* required */
+                            apiKeyRequired: false,
+                        };
+                        apigateway.putMethod(params, function(err, data) {
+                            if (err) {
+                                console.log(err, err.stack);
+                            } 
+                            else {
+                                console.log(data);
+        
+                                // CREATE THE INTEGRATION WITH THE LAMBDA COPY BUCKET FUNCTION
+                                var params = {
+                                    httpMethod: 'POST', /* required */
+                                    integrationHttpMethod: 'POST',
+                                    resourceId: parent_id, /* required */
+                                    restApiId: rest_api_id, /* required */
+                                    type: 'AWS', /* required */
+                                    uri: `arn:aws:apigateway:${process.env.AWS_REGION}:lambda:path/2015-03-31/functions/${arn}/invocations`
+                                };
+                                apigateway.putIntegration(params, function(err, data) {
+                                    if (err) {
+                                        console.log(err, err.stack);
+                                    }
+                                    else {
+                                        console.log(data);
 
-                                            // ASSIGN LAMBDA THE POLICY 
-                                            var params = {
-                                            Action: "lambda:InvokeFunction", 
-                                            FunctionName: funcName, 
-                                            Principal: "apigateway.amazonaws.com", 
-                                            SourceArn: `arn:aws:execute-api:us-east-1:519275522978:${rest_api_id}/*/POST/`,
-                                            StatementId: `ID-${uniqNow}`
-                                           };
-                                           lambda.addPermission(params, function(err, data) {
-                                             if (err) console.log(err, err.stack); // an error occurred
-                                             else     console.log(data);           // successful response
-                                           });
+                                        // ASSIGN POLICY TO THE LAMBDA COPY BUCKET FUNCTION 
+                                        var params = {
+                                        Action: "lambda:InvokeFunction", 
+                                        FunctionName: funcName, 
+                                        Principal: "apigateway.amazonaws.com", 
+                                        SourceArn: `arn:aws:execute-api:us-east-1:519275522978:${rest_api_id}/*/POST/`,
+                                        StatementId: `ID-${uniqNow}`
+                                        };
+                                        lambda.addPermission(params, function(err, data) {
+                                            if (err) console.log(err, err.stack); // an error occurred
+                                            else     console.log(data);           // successful response
+                                        });
 
-                                            //CREATE THE DEPLOYMENT
-                                            var params = {
-                                                restApiId: rest_api_id, /* required */
-                                                description: 'deployment for the REST API for the lambda copy function',
-                                                stageDescription: `stage ${uniqNow} of the REST API for the lambda copy function deployment`,
-                                                stageName: `deployment${uniqNow}`,
-                                            };
-                                            apigateway.createDeployment(params, function(err, data) {
-                                                if (err) {
-                                                    console.log(err, err.stack);
-                                                }
-                                                else {
-                                                    console.log(data);
-                                                    //CREATE PUT METHOD RESPONSE
-                                                    var params = {
-                                                        httpMethod: 'POST', /* required */
-                                                        resourceId: parent_id, /* required */
-                                                        restApiId: rest_api_id, /* required */
-                                                        statusCode: '200', /* required */
-                                                      };
-                                                      apigateway.putMethodResponse(params, function(err, data) {
-                                                        if (err) {
-                                                            console.log(err, err.stack);
-                                                        } 
-                                                        else {
-                                                            console.log(data);
-                                                            //CREATE INTEGRATION RESPONSE
-                                                            var params = {
-                                                                httpMethod: 'POST', /* required */
-                                                                resourceId: parent_id, /* required */
-                                                                restApiId: rest_api_id, /* required */
-                                                                statusCode: '200', /* required */
-                                                              };
-                                                              apigateway.putIntegrationResponse(params, function(err, data) {
-                                                                if (err) console.log(err, err.stack); // an error occurred
-                                                                else     console.log(data);           // successful response
-                                                              });
-
-                                                        }            
-                                                      });
-                                                }
-                                            });
-                                        }     
-                                    });
+                                        // CREATE THE API DEPLOYMENT
+                                        var params = {
+                                            restApiId: rest_api_id, /* required */
+                                            description: 'deployment for the REST API for the lambda copy function',
+                                            stageDescription: `stage ${uniqNow} of the REST API for the lambda copy function deployment`,
+                                            stageName: `deployment${uniqNow}`,
+                                        };
+                                        apigateway.createDeployment(params, function(err, data) {
+                                            if (err) {
+                                                console.log(err, err.stack);
+                                            }
+                                            else {
+                                                console.log(data);
+                                                // CREATE THE METHOD RESPONSE
+                                                var params = {
+                                                    httpMethod: 'POST', /* required */
+                                                    resourceId: parent_id, /* required */
+                                                    restApiId: rest_api_id, /* required */
+                                                    statusCode: '200', /* required */
+                                                    };
+                                                    apigateway.putMethodResponse(params, function(err, data) {
+                                                    if (err) {
+                                                        console.log(err, err.stack);
+                                                    } 
+                                                    else {
+                                                        console.log(data);
+                                                        // CREATE THE INTEGRATION RESPONSE
+                                                        var params = {
+                                                            httpMethod: 'POST', /* required */
+                                                            resourceId: parent_id, /* required */
+                                                            restApiId: rest_api_id, /* required */
+                                                            statusCode: '200', /* required */
+                                                            };
+                                                            apigateway.putIntegrationResponse(params, function(err, data) {
+                                                            if (err) {
+                                                                console.log(err, err.stack);
+                                                            }
+                                                            else {
+                                                                console.log(data);
+                                                            }     
+                                                        });
+                                                    }            
+                                                });
+                                            }
+                                        });
+                                    }     
+                                });
                              
                         }    
                     });   
                 }                
-                });             
+            });             
                        
         }     
-      });
+    });
 }
